debian: No longer build for i386
It was already not possible to use x86_32 hardware because the i386
packages already shipped a 64-bit hypervisor and PV shim. Running 32-bit
utils with a 64-bit hypervisor requires using a compatibility layer that
is fragile and becomes harder to maintain and test upstream.

The libc6-xen package is not being built any more, and 32-bit PV support
also has been removed from the Linux kernel now.

This change ends the 'grace period' in which users should have moved to
using a fully 64-bit dom0.

As a result the only reverse dependency for the libc6-xen package in
Debian is also removed, so this also (Closes: #992909) in the process.

Signed-off-by: Hans van Kranenburg <hans@knorrie.org>