d/control: make xen-hypervisor-common arch specific
It really does not make sense at all to have this package available for
any other arch than that we provide the other packages for. So, do the
same thing like was done for xen-utils-common.

The Debian Policy Manual, in section 5.6.8. Architecture, tells us:
"Specifying a list of architectures or architecture wildcards other than
any is for the minority of cases where a program is not portable or is
not useful on some architectures." Well, in our case, Xen is not
portable to all the other architectures, so this package is not useful
on them.

Signed-off-by: Hans van Kranenburg <hans@knorrie.org>