give meaningful error message if qemu device model is unavailable
There's no sense to switch to qemu-xen-traditional device model
if that one is not enabled in the first place. This way we'll
have a chance later to print a message suggesting to install the
missing qemu package if we *actually* need qemu for the device model.

Gbp-Pq: Name 0022-give-meaningful-error-message-if-qemu-device-model-i.patch