d/*lintian-overrides*: deal with formatting changes
The message output of lintian is changing, and the override lines are
meant to match those messages to silence them.

THEN:
   statically-linked-binary usr/lib/xen-4.16/boot/xen-shim

NOW:
   statically-linked-binary [usr/lib/xen-4.16/boot/xen-shim]

Also see: https://bugs.debian.org/1007002

So, the goal while writing these overrides is not to be as specific as
possible, it's better to be more generic and aim for the lowest chance
to accidentally match (hide) another problem.

For example, for the no-symbols-control-file one in libxenmisc it's the
case for ALL of the so files in the package, so there's not really a
need to put more effort in matching them specifically.

For others, we mostly can just put a * before and after the file names,
so that this will work with old and new lintian versions.

For binary-has-unneeded-section and spelling-error-in-binary, it seems
the order of stuff on the line has changed (filename is now in brackets
and is moved to the beginning of the line), so for them just also use
only the file name with some added *'s... O_o