debian/: Incorporate NMU Part I: The Great Revert
Beause of the time64 transition, NMU's were done to src:xen. In the
packaging repo, we just imported those using dgit.

We have examined the changes and have to do a bunch of cleanup and
housekeeping:
* We have some convenience automation to generate files that have the
  Xen version in them, like fooV.bar.vsn-in -> foo4.17.bar. In the NMU
  package, the input files were lost, and the output was kept.
* The debian/.gitignore file disappeared

The other changes we'll keep, and the way in which we're going to do
that, is to first revert everything, and then apply the relevant parts
again. The result will be that our package repository history will
contain a change/commit that shows the diff of the actual changes,
instead of a very messy and confusing commit that removes unnecessary
parts.