d/{control,rules}: drop support for armhf
During the packaging of Xen 4.19 we discover that it FTBFS on armhf
because OCaml 5.x has no support for native compilation on 32-bit
systems any more.

What we know is that...
* ... Debian bug #989560 exists, which tells us that booting Xen on
  armhf is broken for years already.
* ... we never heard about, or from anyone using this yet, during the
  years we maintain src:xen
* ... the Debian Xen Team does not have the resources and knowledge to
  actively test and support an armhf build, and we never did.

We suspect that not even being able too boot it might just not be the
only thing that would be broken on armhf in Debian.

Instead of trying to build new workarounds into the packaging to barely
rescue it from just not FTBFS-ing, we decide to drop support for the
architecture during the Trixie release cycle.

This also (Closes: #989560) because it's not relevant any more.

Signed-off-by: Hans van Kranenburg <hans@knorrie.org>