debian: undo ABI-stable t64 package renaming
(Comment by Hans van Kranenburg):

The ABI changes caused by switching to 64-bit time_t usage do not affect
the Xen hypervisor and the libraries mentioned in this commit.

These libraries are explicitly engineered to be ABI stable, so they can
be used with any Xen hypervisor version.

Specifically related to time_t, the hypervisor and these libraries use a
`typedef s64 s_time_t`, which can be found in `xen/include/xen/time.h`.

This means for example, that during upgrades, when multiple versions of
Xen can be installed in parallel, there just has to be one version of
these libraries installed.

A different example is that, while we still supported i386 as
toolstack architecture, there was already no 32-bit hypervisor any more,
and the 32-bit tooling was talking to a 64-bit hypervisor already.

All of this does not apply to the 'unstable ABI' libraries, but we ship
all of them inside libxenmiscX.Y. So, libxenmisc4.17 had to be renamed
to libxenmisc4.17t64, but has now again been replaced with
libxenmisc4.19.

Signed-off-by: Maximilian Engelhardt <maxi@daemonizer.de>
Signed-off-by: Hans van Kranenburg <hans@knorrie.org>