debian: ship the new xsd_glue ocaml plugin in xen-utils-V
We ship the new ocaml xsd_glue plugin (domain_getinfo_v1.cmxs) in
xen-utils-V together with oxenstored for now, as it's currently only
used by oxenstored.
If oxenstored gets split out of the xen sources in the future or some
other (external) software starts to use this plugin, we probably need to
update the way we ship ocaml files anyhow, as we currently don't ship
any ocaml development files.

Signed-off-by: Maximilian Engelhardt <maxi@daemonizer.de>